refactor: extract translation languages, favourited by, poll vote, and quote operations into lib/client/statuses.ts
